enums: fixed conversion warning (#4025)

+++ src/tile.h
@@ -138,7 +138,7 @@ class TileItemVector : private ItemVector
 			}
 			return *getBeginDownItem();
 		}
-		void addDownItemCount(int32_t increment) {
+		void addDownItemCount(uint16_t increment) {
 			downItemCount += increment;
 		}
